requirements.txt
langchain
ollama
openai-whisper
torch
torchaudio
numpy
transformers
sentencepiece
ffmpeg-python
